Simulate correlated regressor values

This function simulates regressor values from various marginal distributions with custom correlations.

Usage

correlated_regressors(
  labels,
  n = 100,
  marginals = list(),
  correlation = diag(length(labels)),
  verbose = FALSE
)

Arguments

labels

(character())
Unique labels for the regressors.

n

(integer(1))
The number of values per regressor.

marginals

(list())
Optionally marginal distributions for regressors. If not specified, standard normal marginal distributions are used.

Each list entry must be named according to a regressor label, and the following distributions are currently supported (TODO: if you need another, report here: ...):

discrete distributions

Poisson: list(type = "poisson", lambda = ...)

categorical: list(type = "categorical", p = c(...))

continuous distributions

normal: list(type = "normal", mean = ..., sd = ...)

uniform: list(type = "uniform", min = ..., max = ...)

correlation

(matrix())
A correlation matrix of dimension length(labels), where the (p, q)-th entry defines the correlation between regressor labels[p] and labels[q].

verbose

(logical(1))
Whether to print information about the simulated regressors.

Value

A data.frame with n rows and length(labels) columns.

Details

TODO

Examples

set.seed(1)
labels <- c("P", "C", "N1", "N2", "U")
n <- 100
marginals <- list(
  "P" = list(type = "poisson", lambda = 1),
  "C" = list(type = "categorical", p = c(0.2, 0.3, 0.5)),
  "N1" = list(type = "normal", mean = -1, sd = 2),
  "U" = list(type = "uniform", min = -2, max = -1)
)
correlation <- sample_correlation_matrix(length(labels))
verbose <- TRUE