Generate equidistant vectors in Euclidean space

This function constructs the coordinates of vertices of a regular simplex in \(\mathbb{R}^{\code{dim}}\) and returns the first n of them,

scaled so that the pairwise Euclidean distance between any two vertices equals dist,

and centered so their centroid is at center.

Usage

equidistant_vectors(dim, n = dim + 1, dist = 1, center = rep(0, dim))

Arguments

dim

[integer(1)]
The dimension.

n

[integer(1)]
The number of vertices to return. Cannot be larger than dim + 1.

dist

[numeric(1)]
Desired pairwise Euclidean distance between any two vertices.

center

[numeric(dim)]
Desired center.

Value

A matrix, where each column is a vertex of the simplex.

See also

Examples

dim <- n <- 3
(dist <- runif(1))
#> [1] 0.04928119
(center <- rnorm(dim))
#> [1] -0.3840029  0.7810685 -0.1426429
(V <- equidistant_vectors(dim = dim, n = n, dist = dist, center = center))
#>            [,1]       [,2]       [,3]
#> [1,] -0.3571776 -0.3974155 -0.3974155
#> [2,]  0.7763265  0.8047790  0.7621002
#> [3,] -0.1508564 -0.1508564 -0.1262158
rowMeans(V)
#> [1] -0.3840029  0.7810685 -0.1426429
dist(t(V))
#>            1          2
#> 2 0.04928119           
#> 3 0.04928119 0.04928119