Creating a basic logo for an R package

This function creates a basic R package logo. The logo has a white background and the package name (with or without curly brackets) in the center. The font size for the package name is scaled such that it fits inside the logo. Type ?oeli to see an example.

Usage

package_logo(
  package_name,
  brackets = FALSE,
  background = ggplot2::ggplot() + ggplot2::theme_void(),
  s_x = 1,
  s_y = 1,
  s_width = 1,
  s_height = 1,
  white_around_sticker = FALSE
)

Arguments

package_name

[character(1)]
The package name.

brackets

[logical(1)]
Curly brackets around the package name?

background

A ggplot object, the background of the sticker.

s_x, s_y, s_width, s_height, white_around_sticker

Passed on to sticker.

Value

A ggplot object.

References

This function builds upon sticker.

Use use_logo to set up the logo for a package.

See also

Examples

print(package_logo("my_package", brackets = TRUE))