Using development packages when working with {renv}

This function creates a file that loads development packages so that {renv} can detect and write them to the lockfile.

Usage

renv_development_packages(
  packages = c("covr", "devtools", "DT", "markdown", "R.utils", "yaml"),
  file_name = "development_packages"
)

Arguments

packages

[character()]
Package names.

file_name

[character(1)]
The name for the .R file.

Value

No return value, but it writes a file to the project root and adds an entry to the .Rbuildignore file.

See also