Sample covariance matrix

This function samples a covariance matrix from an inverse Wishart distribution.

Usage

sample_covariance_matrix(dim, df = dim, scale = diag(dim), diag = FALSE)

Arguments

dim

An integer, the dimension.

df

An integer greater or equal dim, the degrees of freedom of the inverse Wishart distribution.

scale

A covariance matrix of dimension dim, the scale matrix of the inverse Wishart distribution.

diag

Set to TRUE for a diagonal matrix.

Value

A covariance matrix.

Examples

sample_covariance_matrix(dim = 3)
#>            [,1]      [,2]       [,3]
#> [1,]  3.7805196 -1.219594  0.9882493
#> [2,] -1.2195937  3.533803 -2.5070029
#> [3,]  0.9882493 -2.507003  1.8405110