Stationary distribution

This function computes the stationary distribution corresponding to a transition probability matrix.

Usage

stationary_distribution(tpm, soft_fail = FALSE)

Arguments

tpm

[matrix()]
A transition probability matrix.

soft_fail

[logical(1)]
Return the discrete uniform distribution if the computation of the stationary distribution fails for some reason? Else, throw an error.

Value

A numeric vector.

See also

Examples

tpm <- matrix(0.05, nrow = 3, ncol = 3)
diag(tpm) <- 0.9
stationary_distribution(tpm)
#> [1] 0.3333333 0.3333333 0.3333333